chore: use user id to identify user's members
