+++ src/features/members/logic/repository/index.ts
@@ -7,7 +7,7 @@ import type { Member, Prisma } from '@prisma/client';
 const prisma = new PrismaClient();
 
 export const createMember = (
-  data: Prisma.MemberCreateInput,
+  data: Prisma.MemberUncheckedCreateInput,
 ): Promise<Result<Member>> => {
   return withResult(() => prisma.member.create({ data }))();
 };
