+++ src/lib/next-auth/index.tsx
@@ -1,0 +1,26 @@
+import { PrismaAdapter } from '@next-auth/prisma-adapter';
+import { PrismaClient } from '@prisma/client';
+import EmailProvider from 'next-auth/providers/email';
+
+import type { AuthOptions } from 'next-auth';
+
+const prisma = new PrismaClient();
+
+export const authOptions: AuthOptions = {
+  adapter: PrismaAdapter(prisma),
+  callbacks: {
+    session: ({ session, user }) => ({
+      ...session,
+      user: {
+        ...user,
+        id: user.id,
+      },
+    }),
+  },
+  providers: [
+    EmailProvider({
+      server: process.env.EMAIL_SERVER,
+      from: process.env.EMAIL_FROM,
+    }),
+  ],
+};
